requirements.txt
fastapi
uvicorn
requests
python-pptx
pollinations
aiosqlite